BibLaTeX configured to only print initials of given names
Until now, we weren't explicitly asking for BibLaTeX to build the citations
with only initial of the author's given names. Therefore when one BibTeX
entry had a full given name and another had only initials, this would also
be reflected in the paper's bibliography.

With this commit, the `giveninits=true' option is given to BibLaTeX to
always only print the initial character of an author's given name.