Distribution tarball now builds in arXiv
`./project make dist' will package all the LaTeX-specific files (and
analysis source files) into one `tar.gz' file that is ready to upload to
servers like arXiv. However, it wasn't updated for some time, so running it
would complain about not having a `configure' script in the top of the
project.

With this commit, it now works with the new file-structure of the project
and also copies all the BibLaTeX source files and `paper.bbl' into the top
tarball directory, which allows arXiv to build the paper as intended.

The output of `./project make dist' has been uploaded and tested on arXiv
and it is built by arXiv perfectly.

Also, a short description of all the special `make' targets was added to
the output of `./project --help'.