Won't copy previous distribution builds in new distribution
Until now, the pipeline was instructed to only ignore the current temporary
project distribution directory. So if there were directories from previous
builds, they were wrongly included in the current tarball.

With this commit, we don't just ignore the directory of the current
distribution, but generally, all directories starting with `paper-v*'.