Possibile to use download-multi-try script without locks
Until now, this script would always only work with a file-lock. But in some
scenarios, we might want to download in parallel. For example when the
system has multiple ports to the internet.

With this commit, we have added this feature: when the lockfile name is
`nolock', it won't lock and will download in parallel.