Corrected markdown code notations in one README.md
In markdown, to specify that part of the text should be shown as code we
need to put it between two `. But in `reproduce/software/make/README.md',
the code parts started with ` and finished in ' (like the file name above).

With this commit, this has been corrected and it will display properly on
web browsers.