Adding X11 libraries path to LDFLAGS in Ghostscript installation
Until this commit, the checking of X11 installation done to ensure that
it is already available in the host system  was crashing in macOS
systems.  The reason is that the place of the X11 libraries use to be
`/opt/X11/lib' in macOS systems.  With this commit, this issue has been
fixed by adding this directory to the LDFLAGS.