Adding basic configuration of Git in README-hacking.md
Until this commit, we were asuming that Git was already properly
configured.  However, in order to be as complete as possible, it would
be good if the basic commands to configure Git were in the
`README-hacking.md'.  With this commit, a small paragraph has been added
in order to have the basic Git configuration commands (i.e. to configure
the name, email, and favorite text editor).