Initial scripts compatible with Dash (minimalistic POSIX)
Until now, the initial project scripts were primarily tested with GNU
Bash. But Bash is not generally available on all systems (it has many
features beyond POSIX). Because of this, effectively we were imposing the
requirement on the user that they must have Bash installed. We recently
started this with setting the shebang of `project' and
`reproduce/software/bash/configure.sh' to `/bin/sh'. After doing so, Raul
and Gaspar reported an error on their systems.

To fix the problem, I installed Dash (a minimalist POSIX-compliant shell)
on my computer and temporarily set the shebangs to `/bin/dash', ran the
project configuration step and fixed all issues that came up. With this
commit, it can go all the way to building GCC on my system's Dash. After
this stage (when `high-level.mk' is called), there is no problem, because
we have our own version of GNU Bash and that installed version is used.
Probably some more issues still remain and will hopefully be found in the
future.

While doing this, I also noticed the following two minor issues:

 - The `./project configure' option `--input-dir' was not recognized
   because it was mistakenly checking `--inputdir'. It has been corrected.

 - The test C programs now use the `<<EOF' method instead of `echo'.

 - In `basic.mk', the extra space between `syspath' and `:=' was removed
   (it was an ancient relic!).