IMPORTANT: reproduce/software/bash renamed to reproduce/software/shell
Until now the shell scripts in the software building phase were in the
`reproduce/software/bash' directory. But given our recent change to a
POSIX-only start, the `configure.sh' shell script (which is the main
component of this directory) is no longer written with Bash.

With this commit, to fix that problem, that directory's name has been
changed to `reproduce/software/shell'.