Better message for analysis when configuration wasn't complete
Until now, when `./project make' was run after an insuccessful run of
`./project configure', it would just say to run `./project configure'. But
for a first time user, this could be confusing because when the
configuration is done in parallel, the error message can be very high on
the command-line outputs and not seen clearly.

With this commit, the error message is more complete and describes the
problem and what the users should do in which circumstance.