Astroquery updated to version 0.4
In the last update of Astropy to version 4.0 they removed some things
that the previous version of Astroquery needs.  As a consequence, it is
also necessary to update the Astroquery version to be a ble to run with
the Astropy 4.0. With this commit, the update of Astroquery to it most
recent version (0.4) has been done.